+++ src/pages/sykmeldt/[oppfolgingsdialogId]/index.tsx
@@ -5,7 +5,10 @@ import { GodkjentPlanAvbrutt } from "../../../components/status/godkjentplanavbr
 import { GodkjennPlanAvslattOgGodkjent } from "../../../components/status/godkjennplanavslattoggodkjent/GodkjennPlanAvslattOgGodkjent";
 import { GodkjennPlanAvslatt } from "../../../components/status/godkjennplanavslatt/GodkjennPlanAvslatt";
 import { OppfolgingsdialogerGodkjenn } from "../../../components/status/godkjennmottatt/OppfolgingsdialogerGodkjenn";
-import { statusPageToDisplaySM } from "utils/statusPageUtils";
+import {
+  getStatusPageTitleAndHeading, StatusPageToDisplay,
+  statusPageToDisplaySM,
+} from "utils/statusPageUtils";
 import { GodkjentPlan } from "../../../components/status/godkjentplan/GodkjentPlan";
 import { IngenPlanTilGodkjenning } from "../../../components/status/ingenplantilgodkjenning/IngenPlanTilGodkjenning";
 import GodkjennPlanSendt from "../../../components/status/godkjennplansendt/GodkjennPlanSendt";
@@ -13,62 +16,13 @@ import { beskyttetSideUtenProps } from "../../../auth/beskyttetSide";
 import SykmeldtSide from "../../../components/blocks/wrappers/SykmeldtSide";
 import { Oppfolgingsplan } from "../../../types/oppfolgingsplan";
 
-export const getStatusPageTitleAndHeading = (
-  oppfolgingsplan?: Oppfolgingsplan
-) => {
-  switch (statusPageToDisplaySM(oppfolgingsplan)) {
-    case "SENDTPLANTILGODKJENNING": {
-      return {
-        title: `Status på oppfølgingsplan`,
-        heading: `Sendt til godkjenning`,
-      };
-    }
-    case "MOTTATTFLEREGODKJENNINGER": {
-      return {
-        title: `Status på oppfølgingsplan`,
-        heading: `Mottatt endring`,
-      };
-    }
-    case "GODKJENNPLANMOTTATT": {
-      return {
-        title: `Godkjenn ${oppfolgingsplan?.virksomhet?.navn}`,
-        heading: `Du har mottatt en ny plan for ${oppfolgingsplan?.virksomhet?.navn}`,
-      };
-    }
-    case "GODKJENNPLANAVSLATT": {
-      return {
-        title: `Status på oppfølgingsplan`,
-        heading: `Lederen din har noen forslag`,
-      };
-    }
-    case "GODKJENTPLANAVBRUTT": {
-      return {
-        title: `Status på oppfølgingsplan`,
-        heading: `Tidligere oppfølgingsplan`,
-      };
-    }
-    case "GODKJENTPLAN": {
-      return {
-        title: `Status på oppfølgingsplan`,
-        heading: `Oppfølgingsplan`,
-      };
-    }
-    default: {
-      return {
-        title: `Status på oppfølgingsplan`,
-        heading: `Status på oppfølgingsplan`,
-      };
-    }
-  }
-};
-
 interface ContentProps {
   oppfolgingsplan?: Oppfolgingsplan;
+  pageToDisplay: StatusPageToDisplay | null;
 }
 
-const Content = ({ oppfolgingsplan }: ContentProps): ReactElement | null => {
+const Content = ({ oppfolgingsplan, pageToDisplay }: ContentProps): ReactElement | null => {
   if (!oppfolgingsplan) return null;
-  const pageToDisplay = statusPageToDisplaySM(oppfolgingsplan);
 
   switch (pageToDisplay) {
     case "SENDTPLANTILGODKJENNING": {
@@ -99,11 +53,17 @@ const Content = ({ oppfolgingsplan }: ContentProps): ReactElement | null => {
 
 const OppfolgingsplanStatus: NextPage = () => {
   const aktivPlan = useAktivPlanSM();
-  const { title, heading } = getStatusPageTitleAndHeading(aktivPlan);
+
+  const pageToDisplay = statusPageToDisplaySM(aktivPlan);
+  const { title, heading } = getStatusPageTitleAndHeading(
+    pageToDisplay,
+    aktivPlan?.virksomhet?.navn,
+    "Lederen din"
+  );
 
   return (
     <SykmeldtSide title={title} heading={heading}>
-      <Content oppfolgingsplan={aktivPlan} />
+      <Content oppfolgingsplan={aktivPlan} pageToDisplay={pageToDisplay} />
     </SykmeldtSide>
   );
 };
