+++ src/server/data/mock/getMockDb.ts
@@ -15,9 +15,9 @@ export type TestScenario =
   | "INGENPLAN"
   | "TIDLIGEREPLANER"
   | "UNDERARBEID"
-  | "GODKJENNPLANAVSLATT"
-  | "GODKJENNPLANSENDT"
-  | "GODKJENNPLANMOTTATT";
+  | "SYKMELDT_SENDT_TIL_GODKJENNING_AG_HAR_ENDRET"
+  | "SYKMELDT_SENDT_TIL_GODKJENNING"
+  | "ARBEIDSGIVER_SENDT_TIL_GODKJENNING";
 
 export interface MockSetup {
   oppfolgingsplaner: OppfolgingsplanDTO[];
