--- src/server/data/mock/testscenarios/godkjennPlanSendt/godkjennPlanSendtScenario.ts
@@ -1,9 +1,0 @@
-import { godkjennPlanSendtOppfolgingsplanerMockData } from "./godkjennPlanSendtOppfolgingsplanerMockData";
-import { MockSetup } from "../../getMockDb";
-import { defaultMockSetup } from "../../defaultData/defaultMockSetup";
-
-export const godkjennPlanSendtScenario: MockSetup = {
-  ...defaultMockSetup,
-  oppfolgingsplaner: godkjennPlanSendtOppfolgingsplanerMockData,
-  activeTestScenario: "GODKJENNPLANSENDT",
-};
