--- src/server/data/mock/testscenarios/godkjentPlanAvslatt/godkjennPlanAvslattOppfolgingsplanerMockData.ts
@@ -1,1923 +1,0 @@
-import { leggTilDagerPaDato } from "utils/dateUtils";
-import { defaultNarmesteLedereMockData } from "server/data/mock/defaultData/oppfolgingsplanservice/defaultNarmesteLedereMockData";
-import { OppfolgingsplanDTO } from "../../../../../schema/oppfolgingsplanSchema";
-
-const today = new Date();
-
-export const godkjennPlanAvslattOppfolgingsplanerMockData: OppfolgingsplanDTO[] =
-  [
-    {
-      id: 3455,
-      sistEndretDato: "2020-02-20T11:31:29.837",
-      opprettetDato: "2020-02-20",
-      status: "UNDER_ARBEID",
-      virksomhet: {
-        virksomhetsnummer: "110110110",
-        navn: "",
-      },
-      godkjentPlan: null,
-      godkjenninger: [
-        {
-          godkjent: false,
-          godkjentAv: {
-            navn: "Albus Dumbledore",
-            fnr: "11011011011",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          beskrivelse: "Beskrivelse",
-          godkjenningsTidspunkt: "2022-10-10T10:46:06.041",
-          gyldighetstidspunkt: {
-            fom: "2022-10-20T10:46:06.041",
-            tom: "2022-10-29T10:46:06.041",
-            evalueres: "2022-11-01T10:46:06.041",
-          },
-          delMedNav: false,
-        },
-      ],
-      arbeidsoppgaveListe: [
-        {
-          arbeidsoppgaveId: 6406,
-          arbeidsoppgavenavn: "Kontorarbeid ",
-          erVurdertAvSykmeldt: true,
-          gjennomfoering: {
-            kanGjennomfoeres: "TILRETTELEGGING",
-            paaAnnetSted: true,
-            medMerTid: true,
-            medHjelp: false,
-            kanBeskrivelse:
-              "Trenger mer tid til de faste oppgavene og kan utføre noen av disse hjemmefra. ",
-            kanIkkeBeskrivelse: null,
-          },
-          opprettetDato: "2020-02-20T11:31:30.305",
-          sistEndretDato: "2020-02-20T11:31:30.305",
-          sistEndretAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          opprettetAv: {
-            navn: "Labben Krabbeklo",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-        {
-          arbeidsoppgaveId: 6407,
-          arbeidsoppgavenavn: "Utviklingssamtaler",
-          erVurdertAvSykmeldt: true,
-          gjennomfoering: {
-            kanGjennomfoeres: "KAN",
-            paaAnnetSted: null,
-            medMerTid: null,
-            medHjelp: null,
-            kanBeskrivelse: null,
-            kanIkkeBeskrivelse: null,
-          },
-          opprettetDato: "2020-02-20T11:31:30.318",
-          sistEndretDato: "2020-02-20T11:31:30.318",
-          sistEndretAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          opprettetAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-        {
-          arbeidsoppgaveId: 6408,
-          arbeidsoppgavenavn: "Foreldremøte ",
-          erVurdertAvSykmeldt: true,
-          gjennomfoering: {
-            kanGjennomfoeres: "KAN_IKKE",
-            paaAnnetSted: null,
-            medMerTid: null,
-            medHjelp: null,
-            kanBeskrivelse: null,
-            kanIkkeBeskrivelse: "Blir for krevende ",
-          },
-          opprettetDato: "2020-02-20T11:31:30.319",
-          sistEndretDato: "2020-02-20T11:31:30.319",
-          sistEndretAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          opprettetAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-        {
-          arbeidsoppgaveId: 6409,
-          arbeidsoppgavenavn: "Undervisning ",
-          erVurdertAvSykmeldt: false,
-          gjennomfoering: null,
-          opprettetDato: "2020-02-20T11:31:30.322",
-          sistEndretDato: "2020-02-20T11:31:30.322",
-          sistEndretAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          opprettetAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-      ],
-      tiltakListe: [
-        {
-          tiltakId: 4852,
-          tiltaknavn: "Hjemmekontor ",
-          knyttetTilArbeidsoppgaveId: null,
-          fom: "2021-04-15",
-          tom: "2022-07-26",
-          beskrivelse:
-            "Lage og rette prøver, forberede utviklingssamtaler, samt andre enklere oppgaver ",
-          beskrivelseIkkeAktuelt: null,
-          opprettetDato: "2020-02-20T11:31:30.338",
-          sistEndretDato: "2020-02-20T11:31:30.338",
-          kommentarer: [
-            {
-              id: 782,
-              tekst:
-                "Her må vi bli enige om antall timer og hvor mange dager. ",
-              opprettetTidspunkt: "2020-02-20T11:31:30.352",
-              sistEndretDato: "2020-02-20T11:31:30.352",
-              opprettetAv: {
-                navn: defaultNarmesteLedereMockData[0].navn,
-                fnr: defaultNarmesteLedereMockData[0].fnr,
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-              sistEndretAv: {
-                navn: defaultNarmesteLedereMockData[0].navn,
-                fnr: defaultNarmesteLedereMockData[0].fnr,
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-            },
-          ],
-          status: "FORSLAG",
-          gjennomfoering: null,
-          opprettetAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          sistEndretAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-        {
-          tiltakId: 4853,
-          tiltaknavn: "Kortere arbeidsdag",
-          knyttetTilArbeidsoppgaveId: null,
-          fom: "2020-02-10",
-          tom: "2020-02-21",
-          beskrivelse: "Undervise i ett fag for å unngå overbelastning ",
-          beskrivelseIkkeAktuelt: null,
-          opprettetDato: "2020-02-20T11:31:30.373",
-          sistEndretDato: "2020-02-20T11:31:30.373",
-          kommentarer: [
-            {
-              id: 783,
-              tekst: "Hvilket fag passer det best å starte med? ",
-              opprettetTidspunkt: "2020-02-20T11:31:30.376",
-              sistEndretDato: "2020-02-20T11:31:30.376",
-              opprettetAv: {
-                navn: defaultNarmesteLedereMockData[0].navn,
-                fnr: defaultNarmesteLedereMockData[0].fnr,
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-              sistEndretAv: {
-                navn: defaultNarmesteLedereMockData[0].navn,
-                fnr: defaultNarmesteLedereMockData[0].fnr,
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-            },
-            {
-              id: 784,
-              tekst: "Gjerne KRLE eller samfunnsfag.",
-              opprettetTidspunkt: "2020-02-20T11:31:30.378",
-              sistEndretDato: "2020-02-20T11:31:30.378",
-              opprettetAv: {
-                navn: "Labben Rekemester",
-                fnr: "110110110110",
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-              sistEndretAv: {
-                navn: "Labben Rekemester",
-                fnr: "110110110110",
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-            },
-            {
-              id: 785,
-              tekst: "Men, ønsker kun å undervise digitalt.",
-              opprettetTidspunkt: "2020-02-21T12:33:12.378",
-              sistEndretDato: "2020-02-21T12:33:12.378",
-              opprettetAv: {
-                navn: "Labben Rekemester",
-                fnr: "110110110110",
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-              sistEndretAv: {
-                navn: "Labben Rekemester",
-                fnr: "110110110110",
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-            },
-          ],
-          status: "AVTALT",
-          gjennomfoering:
-            "Vi ser hvordan det går denne perioden og øker gradvis hvis aktuelt. ",
-          opprettetAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          sistEndretAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-        {
-          tiltakId: 4854,
-          tiltaknavn: "Assistent i klasserommet ",
-          knyttetTilArbeidsoppgaveId: null,
-          fom: null,
-          tom: null,
-          beskrivelse: "Assistent til å bistå i undervisningen.",
-          beskrivelseIkkeAktuelt:
-            "Skolen har ikke ressurser til å kunne tilby dette. ",
-          opprettetDato: "2020-02-20T11:31:30.38",
-          sistEndretDato: "2020-02-20T11:31:30.38",
-          kommentarer: [],
-          status: "IKKE_AKTUELT",
-          gjennomfoering: null,
-          opprettetAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          sistEndretAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-      ],
-      avbruttPlanListe: [
-        {
-          tidspunkt: "2020-02-20T11:31:29.835",
-          id: 3451,
-        },
-        {
-          tidspunkt: "2020-02-12T15:24:53.34",
-          id: 3450,
-        },
-      ],
-      arbeidsgiver: {
-        narmesteLeder: {
-          navn: "Labben Rekemester",
-          fnr: "20405060718",
-          epost: null,
-          tlf: null,
-          sistInnlogget: "2020-02-24T08:57:46.747",
-          samtykke: null,
-          virksomhetsnummer: "123",
-          erAktiv: false,
-          aktivFom: "2020-02-24T08:57:46.747",
-          aktivTom: null,
-        },
-      },
-      arbeidstaker: {
-        navn: "Labben Rekemester",
-        fnr: "110110110112",
-        epost: null,
-        tlf: null,
-        sistInnlogget: "2020-02-24T09:34:24.266",
-        samtykke: null,
-        evaluering: null,
-        stillinger: [
-          {
-            virksomhetsnummer: "110110110",
-            yrke: "Soppkontrollør",
-            prosent: 80,
-            fom: "2000-02-11T15:34:53.787",
-            tom: "2200-02-11T15:34:53.787",
-          },
-        ],
-      },
-      sistEndretAv: {
-        navn: defaultNarmesteLedereMockData[0].navn,
-        fnr: defaultNarmesteLedereMockData[0].fnr,
-        epost: null,
-        tlf: null,
-        sistInnlogget: null,
-        samtykke: null,
-        evaluering: null,
-        stillinger: [],
-      },
-    },
-    {
-      id: 6456,
-      sistEndretDato: "2020-02-20T11:31:29.837",
-      opprettetDato: "2020-02-20",
-      status: "AKTIV",
-      virksomhet: {
-        virksomhetsnummer: "123456789",
-        navn: "",
-      },
-      godkjentPlan: {
-        opprettetTidspunkt: leggTilDagerPaDato(today, -7).toJSON(),
-        gyldighetstidspunkt: {
-          fom: leggTilDagerPaDato(today, -7).toJSON(),
-          tom: leggTilDagerPaDato(today, 7).toJSON(),
-          // fom: "2019-05-10T11:31:29.837",
-          // tom: "2019-07-15T11:31:29.837",
-          evalueres: leggTilDagerPaDato(today, 14).toJSON(),
-        },
-        tvungenGodkjenning: true,
-        deltMedNAVTidspunkt: null,
-        deltMedNAV: false,
-        deltMedFastlegeTidspunkt: null,
-        deltMedFastlege: false,
-        dokumentUuid: "12345678-1234-1234-1234-123456789abc",
-        avbruttPlan: null,
-      },
-      godkjenninger: [],
-      arbeidsoppgaveListe: [
-        {
-          arbeidsoppgaveId: 6406,
-          arbeidsoppgavenavn: "Kontorarbeid ",
-          erVurdertAvSykmeldt: true,
-          gjennomfoering: {
-            kanGjennomfoeres: "TILRETTELEGGING",
-            paaAnnetSted: true,
-            medMerTid: true,
-            medHjelp: true,
-            kanBeskrivelse:
-              "Trenger mer tid til de faste oppgavene og kan utføre noen av disse hjemmefra. ",
-            kanIkkeBeskrivelse: null,
-          },
-          opprettetDato: "2020-02-20T11:31:30.305",
-          sistEndretDato: "2020-02-20T11:31:30.305",
-          sistEndretAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          opprettetAv: {
-            navn: "Labben Krabbeklo",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-        {
-          arbeidsoppgaveId: 6407,
-          arbeidsoppgavenavn: "Utviklingssamtaler",
-          erVurdertAvSykmeldt: true,
-          gjennomfoering: {
-            kanGjennomfoeres: "KAN",
-            paaAnnetSted: null,
-            medMerTid: null,
-            medHjelp: null,
-            kanBeskrivelse: null,
-            kanIkkeBeskrivelse: null,
-          },
-          opprettetDato: "2020-02-20T11:31:30.318",
-          sistEndretDato: "2020-02-20T11:31:30.318",
-          sistEndretAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          opprettetAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-        {
-          arbeidsoppgaveId: 6408,
-          arbeidsoppgavenavn: "Foreldremøte ",
-          erVurdertAvSykmeldt: true,
-          gjennomfoering: {
-            kanGjennomfoeres: "KAN_IKKE",
-            paaAnnetSted: null,
-            medMerTid: null,
-            medHjelp: null,
-            kanBeskrivelse: null,
-            kanIkkeBeskrivelse: "Blir for krevende ",
-          },
-          opprettetDato: "2020-02-20T11:31:30.319",
-          sistEndretDato: "2020-02-20T11:31:30.319",
-          sistEndretAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          opprettetAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-        {
-          arbeidsoppgaveId: 6409,
-          arbeidsoppgavenavn: "Undervisning ",
-          erVurdertAvSykmeldt: false,
-          gjennomfoering: null,
-          opprettetDato: "2020-02-20T11:31:30.322",
-          sistEndretDato: "2020-02-20T11:31:30.322",
-          sistEndretAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          opprettetAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-      ],
-      tiltakListe: [
-        {
-          tiltakId: 4852,
-          tiltaknavn: "Hjemmekontor ",
-          knyttetTilArbeidsoppgaveId: null,
-          fom: "2020-02-10",
-          tom: "2020-02-21",
-          beskrivelse:
-            "Lage og rette prøver, forberede utviklingssamtaler, samt andre enklere oppgaver ",
-          beskrivelseIkkeAktuelt: null,
-          opprettetDato: "2020-02-20T11:31:30.338",
-          sistEndretDato: "2020-02-20T11:31:30.338",
-          kommentarer: [
-            {
-              id: 782,
-              tekst:
-                "Her må vi bli enige om antall timer og hvor mange dager. ",
-              opprettetTidspunkt: "2020-02-20T11:31:30.352",
-              sistEndretDato: "2020-02-20T11:31:30.352",
-              opprettetAv: {
-                navn: defaultNarmesteLedereMockData[0].navn,
-                fnr: defaultNarmesteLedereMockData[0].fnr,
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-              sistEndretAv: {
-                navn: defaultNarmesteLedereMockData[0].navn,
-                fnr: defaultNarmesteLedereMockData[0].fnr,
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-            },
-          ],
-          status: "FORSLAG",
-          gjennomfoering: null,
-          opprettetAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          sistEndretAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-        {
-          tiltakId: 4853,
-          tiltaknavn: "Kortere arbeidsdag",
-          knyttetTilArbeidsoppgaveId: null,
-          fom: "2020-02-10",
-          tom: "2020-02-21",
-          beskrivelse: "Undervise i ett fag for å unngå overbelastning ",
-          beskrivelseIkkeAktuelt: null,
-          opprettetDato: "2020-02-20T11:31:30.373",
-          sistEndretDato: "2020-02-20T11:31:30.373",
-          kommentarer: [
-            {
-              id: 783,
-              tekst: "Hvilket fag passer det best å starte med? ",
-              opprettetTidspunkt: "2020-02-20T11:31:30.376",
-              sistEndretDato: "2020-02-20T11:31:30.376",
-              opprettetAv: {
-                navn: defaultNarmesteLedereMockData[0].navn,
-                fnr: defaultNarmesteLedereMockData[0].fnr,
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-              sistEndretAv: {
-                navn: defaultNarmesteLedereMockData[0].navn,
-                fnr: defaultNarmesteLedereMockData[0].fnr,
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-            },
-            {
-              id: 784,
-              tekst: "Gjerne KRLE eller samfunnsfag.",
-              opprettetTidspunkt: "2020-02-20T11:31:30.378",
-              sistEndretDato: "2020-02-20T11:31:30.378",
-              opprettetAv: {
-                navn: "Labben Rekemester",
-                fnr: "110110110110",
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-              sistEndretAv: {
-                navn: "Labben Rekemester",
-                fnr: "110110110110",
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-            },
-            {
-              id: 785,
-              tekst: "Men, ønsker kun å undervise digitalt.",
-              opprettetTidspunkt: "2020-02-21T12:33:12.378",
-              sistEndretDato: "2020-02-21T12:33:12.378",
-              opprettetAv: {
-                navn: "Labben Rekemester",
-                fnr: "110110110110",
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-              sistEndretAv: {
-                navn: "Labben Rekemester",
-                fnr: "110110110110",
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-            },
-          ],
-          status: "AVTALT",
-          gjennomfoering:
-            "Vi ser hvordan det går denne perioden og øker gradvis hvis aktuelt. ",
-          opprettetAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          sistEndretAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-        {
-          tiltakId: 4854,
-          tiltaknavn: "Assistent i klasserommet ",
-          knyttetTilArbeidsoppgaveId: null,
-          fom: null,
-          tom: null,
-          beskrivelse: "Assistent til å bistå i undervisningen.",
-          beskrivelseIkkeAktuelt:
-            "Skolen har ikke ressurser til å kunne tilby dette. ",
-          opprettetDato: "2020-02-20T11:31:30.38",
-          sistEndretDato: "2020-02-20T11:31:30.38",
-          kommentarer: [],
-          status: "IKKE_AKTUELT",
-          gjennomfoering: null,
-          opprettetAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          sistEndretAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-      ],
-      avbruttPlanListe: [
-        {
-          tidspunkt: "2020-02-20T11:31:29.835",
-          id: 3451,
-        },
-        {
-          tidspunkt: "2020-02-12T15:24:53.34",
-          id: 3450,
-        },
-      ],
-      arbeidsgiver: {
-        narmesteLeder: {
-          navn: "Albus Dumbledore",
-          fnr: "11011011011",
-          epost: "albus@hogwarts.ac.uk",
-          tlf: "110",
-          sistInnlogget: "2020-02-24T08:57:46.747",
-          samtykke: null,
-          virksomhetsnummer: "110110110",
-          erAktiv: true,
-          aktivFom: "2020-02-24T08:57:46.747",
-          aktivTom: null,
-        },
-      },
-      arbeidstaker: {
-        navn: "Labben Rekemester",
-        fnr: "110110110110",
-        epost: null,
-        tlf: null,
-        sistInnlogget: "2020-02-24T09:34:24.266",
-        samtykke: null,
-        evaluering: null,
-        stillinger: [
-          {
-            virksomhetsnummer: "110110110",
-            yrke: "Soppkontrollør",
-            prosent: 80,
-            fom: "2000-02-11T15:34:53.787",
-            tom: "2200-02-11T15:34:53.787",
-          },
-        ],
-      },
-      sistEndretAv: {
-        navn: defaultNarmesteLedereMockData[1].navn,
-        fnr: defaultNarmesteLedereMockData[1].fnr,
-        epost: null,
-        tlf: null,
-        sistInnlogget: null,
-        samtykke: null,
-        evaluering: null,
-        stillinger: [],
-      },
-    },
-    {
-      id: 7777,
-      sistEndretDato: "2020-02-20T11:31:29.837",
-      opprettetDato: "2020-02-20",
-      status: "UTDATERT",
-      virksomhet: {
-        virksomhetsnummer: "110110110",
-        navn: "",
-      },
-      godkjentPlan: {
-        opprettetTidspunkt: leggTilDagerPaDato(today, -7).toJSON(),
-        gyldighetstidspunkt: {
-          fom: "2015-09-11T11:31:29.837",
-          tom: "2015-11-16T11:31:29.837",
-          evalueres: leggTilDagerPaDato(today, 14).toJSON(),
-        },
-        tvungenGodkjenning: true,
-        deltMedNAVTidspunkt: null,
-        deltMedNAV: false,
-        deltMedFastlegeTidspunkt: null,
-        deltMedFastlege: false,
-        dokumentUuid: "12345678-1234-1234-1234-123456789abc",
-        avbruttPlan: null,
-      },
-      godkjenninger: [
-        {
-          godkjent: true,
-          godkjentAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          beskrivelse: "Beskrivelse",
-          godkjenningsTidspunkt: "2022-10-10T10:46:06.041",
-          gyldighetstidspunkt: {
-            fom: "2022-10-20T10:46:06.041",
-            tom: "2022-10-29T10:46:06.041",
-            evalueres: "2022-11-01T10:46:06.041",
-          },
-          delMedNav: false,
-        },
-      ],
-      arbeidsoppgaveListe: [
-        {
-          arbeidsoppgaveId: 6406,
-          arbeidsoppgavenavn: "Kontorarbeid ",
-          erVurdertAvSykmeldt: true,
-          gjennomfoering: {
-            kanGjennomfoeres: "TILRETTELEGGING",
-            paaAnnetSted: true,
-            medMerTid: true,
-            medHjelp: true,
-            kanBeskrivelse:
-              "Trenger mer tid til de faste oppgavene og kan utføre noen av disse hjemmefra. ",
-            kanIkkeBeskrivelse: null,
-          },
-          opprettetDato: "2020-02-20T11:31:30.305",
-          sistEndretDato: "2020-02-20T11:31:30.305",
-          sistEndretAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          opprettetAv: {
-            navn: "Labben Krabbeklo",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-        {
-          arbeidsoppgaveId: 6407,
-          arbeidsoppgavenavn: "Utviklingssamtaler",
-          erVurdertAvSykmeldt: true,
-          gjennomfoering: {
-            kanGjennomfoeres: "KAN",
-            paaAnnetSted: null,
-            medMerTid: null,
-            medHjelp: null,
-            kanBeskrivelse: null,
-            kanIkkeBeskrivelse: null,
-          },
-          opprettetDato: "2020-02-20T11:31:30.318",
-          sistEndretDato: "2020-02-20T11:31:30.318",
-          sistEndretAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          opprettetAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-        {
-          arbeidsoppgaveId: 6408,
-          arbeidsoppgavenavn: "Foreldremøte ",
-          erVurdertAvSykmeldt: true,
-          gjennomfoering: {
-            kanGjennomfoeres: "KAN_IKKE",
-            paaAnnetSted: null,
-            medMerTid: null,
-            medHjelp: null,
-            kanBeskrivelse: null,
-            kanIkkeBeskrivelse: "Blir for krevende ",
-          },
-          opprettetDato: "2020-02-20T11:31:30.319",
-          sistEndretDato: "2020-02-20T11:31:30.319",
-          sistEndretAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          opprettetAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-        {
-          arbeidsoppgaveId: 6409,
-          arbeidsoppgavenavn: "Undervisning ",
-          erVurdertAvSykmeldt: false,
-          gjennomfoering: null,
-          opprettetDato: "2020-02-20T11:31:30.322",
-          sistEndretDato: "2020-02-20T11:31:30.322",
-          sistEndretAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          opprettetAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-      ],
-      tiltakListe: [
-        {
-          tiltakId: 4852,
-          tiltaknavn: "Hjemmekontor ",
-          knyttetTilArbeidsoppgaveId: null,
-          fom: "2020-02-10",
-          tom: "2020-02-21",
-          beskrivelse:
-            "Lage og rette prøver, forberede utviklingssamtaler, samt andre enklere oppgaver ",
-          beskrivelseIkkeAktuelt: null,
-          opprettetDato: "2020-02-20T11:31:30.338",
-          sistEndretDato: "2020-02-20T11:31:30.338",
-          kommentarer: [
-            {
-              id: 782,
-              tekst:
-                "Her må vi bli enige om antall timer og hvor mange dager. ",
-              opprettetTidspunkt: "2020-02-20T11:31:30.352",
-              sistEndretDato: "2020-02-20T11:31:30.352",
-              opprettetAv: {
-                navn: defaultNarmesteLedereMockData[0].navn,
-                fnr: defaultNarmesteLedereMockData[0].fnr,
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-              sistEndretAv: {
-                navn: defaultNarmesteLedereMockData[0].navn,
-                fnr: defaultNarmesteLedereMockData[0].fnr,
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-            },
-          ],
-          status: "FORSLAG",
-          gjennomfoering: null,
-          opprettetAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          sistEndretAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-        {
-          tiltakId: 4853,
-          tiltaknavn: "Kortere arbeidsdag",
-          knyttetTilArbeidsoppgaveId: null,
-          fom: "2020-02-10",
-          tom: "2020-02-21",
-          beskrivelse: "Undervise i ett fag for å unngå overbelastning ",
-          beskrivelseIkkeAktuelt: null,
-          opprettetDato: "2020-02-20T11:31:30.373",
-          sistEndretDato: "2020-02-20T11:31:30.373",
-          kommentarer: [
-            {
-              id: 783,
-              tekst: "Hvilket fag passer det best å starte med? ",
-              opprettetTidspunkt: "2020-02-20T11:31:30.376",
-              sistEndretDato: "2020-02-20T11:31:30.376",
-              opprettetAv: {
-                navn: defaultNarmesteLedereMockData[0].navn,
-                fnr: defaultNarmesteLedereMockData[0].fnr,
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-              sistEndretAv: {
-                navn: defaultNarmesteLedereMockData[0].navn,
-                fnr: defaultNarmesteLedereMockData[0].fnr,
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-            },
-            {
-              id: 784,
-              tekst: "Gjerne KRLE eller samfunnsfag.",
-              opprettetTidspunkt: "2020-02-20T11:31:30.378",
-              sistEndretDato: "2020-02-20T11:31:30.378",
-              opprettetAv: {
-                navn: "Labben Rekemester",
-                fnr: "110110110110",
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-              sistEndretAv: {
-                navn: "Labben Rekemester",
-                fnr: "110110110110",
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-            },
-            {
-              id: 785,
-              tekst: "Men, ønsker kun å undervise digitalt.",
-              opprettetTidspunkt: "2020-02-21T12:33:12.378",
-              sistEndretDato: "2020-02-21T12:33:12.378",
-              opprettetAv: {
-                navn: "Labben Rekemester",
-                fnr: "110110110110",
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-              sistEndretAv: {
-                navn: "Labben Rekemester",
-                fnr: "110110110110",
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-            },
-          ],
-          status: "AVTALT",
-          gjennomfoering:
-            "Vi ser hvordan det går denne perioden og øker gradvis hvis aktuelt. ",
-          opprettetAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          sistEndretAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-        {
-          tiltakId: 4854,
-          tiltaknavn: "Assistent i klasserommet ",
-          knyttetTilArbeidsoppgaveId: null,
-          fom: null,
-          tom: null,
-          beskrivelse: "Assistent til å bistå i undervisningen.",
-          beskrivelseIkkeAktuelt:
-            "Skolen har ikke ressurser til å kunne tilby dette. ",
-          opprettetDato: "2020-02-20T11:31:30.38",
-          sistEndretDato: "2020-02-20T11:31:30.38",
-          kommentarer: [],
-          status: "IKKE_AKTUELT",
-          gjennomfoering: null,
-          opprettetAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          sistEndretAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-      ],
-      avbruttPlanListe: [
-        {
-          tidspunkt: "2020-02-20T11:31:29.835",
-          id: 3451,
-        },
-        {
-          tidspunkt: "2020-02-12T15:24:53.34",
-          id: 3450,
-        },
-      ],
-      arbeidsgiver: {
-        narmesteLeder: {
-          navn: "Albus Dumbledore",
-          fnr: "11011011011",
-          epost: "albus@hogwarts.ac.uk",
-          tlf: "110",
-          sistInnlogget: "2020-02-24T08:57:46.747",
-          samtykke: null,
-          virksomhetsnummer: "110110110",
-          erAktiv: true,
-          aktivFom: "2020-02-24T08:57:46.747",
-          aktivTom: null,
-        },
-      },
-      arbeidstaker: {
-        navn: "Labben Rekemester",
-        fnr: "110110110110",
-        epost: null,
-        tlf: null,
-        sistInnlogget: "2020-02-24T09:34:24.266",
-        samtykke: null,
-        evaluering: null,
-        stillinger: [
-          {
-            virksomhetsnummer: "110110110",
-            yrke: "Soppkontrollør",
-            prosent: 80,
-            fom: "2000-02-11T15:34:53.787",
-            tom: "2200-02-11T15:34:53.787",
-          },
-        ],
-      },
-      sistEndretAv: {
-        navn: defaultNarmesteLedereMockData[0].navn,
-        fnr: defaultNarmesteLedereMockData[0].fnr,
-        epost: null,
-        tlf: null,
-        sistInnlogget: null,
-        samtykke: null,
-        evaluering: null,
-        stillinger: [],
-      },
-    },
-    {
-      id: 3450,
-      sistEndretDato: "2020-02-11T15:34:53.787",
-      opprettetDato: "2020-02-11",
-      status: "AVBRUTT",
-      virksomhet: {
-        virksomhetsnummer: "110110110",
-        navn: "",
-      },
-      godkjentPlan: {
-        opprettetTidspunkt: "2020-02-11T15:34:52.311",
-        gyldighetstidspunkt: {
-          fom: "2020-02-11",
-          tom: "2020-02-23",
-          evalueres: "2020-02-28",
-        },
-        tvungenGodkjenning: false,
-        deltMedNAVTidspunkt: null,
-        deltMedNAV: false,
-        deltMedFastlegeTidspunkt: null,
-        deltMedFastlege: false,
-        dokumentUuid: "ceed5fe6-7149-4a77-9189-f8ef055ce5d2",
-        avbruttPlan: {
-          tidspunkt: "2020-02-12T15:24:53.34",
-          id: 3450,
-        },
-      },
-      godkjenninger: [],
-      arbeidsoppgaveListe: [
-        {
-          arbeidsoppgaveId: 6384,
-          arbeidsoppgavenavn: "Dette er en fin arbeidsoppgave",
-          erVurdertAvSykmeldt: true,
-          gjennomfoering: {
-            kanGjennomfoeres: "KAN",
-            paaAnnetSted: null,
-            medMerTid: null,
-            medHjelp: null,
-            kanBeskrivelse: null,
-            kanIkkeBeskrivelse: null,
-          },
-          opprettetDato: "2020-02-11T15:13:58.281",
-          sistEndretDato: "2020-02-11T15:13:58.281",
-          sistEndretAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          opprettetAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-        {
-          arbeidsoppgaveId: 6382,
-          arbeidsoppgavenavn: "Wretrwtryhujk",
-          erVurdertAvSykmeldt: true,
-          gjennomfoering: {
-            kanGjennomfoeres: "TILRETTELEGGING",
-            paaAnnetSted: false,
-            medMerTid: false,
-            medHjelp: true,
-            kanBeskrivelse:
-              "akhdhadhka kjahdkjah dhakjdh. hkajhdk akd. hkjhj. ",
-            kanIkkeBeskrivelse: null,
-          },
-          opprettetDato: "2020-02-11T14:57:37.708",
-          sistEndretDato: "2020-02-11T15:13:09.211",
-          sistEndretAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          opprettetAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-      ],
-      tiltakListe: [
-        {
-          tiltakId: 4827,
-          tiltaknavn: "Dette er et tiltak som skal gjøres på arbeidsplassen",
-          knyttetTilArbeidsoppgaveId: null,
-          fom: "2020-02-11",
-          tom: "2020-02-23",
-          beskrivelse: "Et stort tiltak",
-          beskrivelseIkkeAktuelt: null,
-          opprettetDato: "2020-02-11T15:23:36.122",
-          sistEndretDato: "2020-02-11T15:23:36.122",
-          kommentarer: [],
-          status: "FORSLAG",
-          gjennomfoering: null,
-          opprettetAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          sistEndretAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-      ],
-      avbruttPlanListe: [],
-      arbeidsgiver: {
-        narmesteLeder: {
-          navn: "Labben Rekemester",
-          fnr: "20405060718",
-          epost: null,
-          tlf: null,
-          sistInnlogget: "2020-02-12T15:24:54.295",
-          samtykke: true,
-          virksomhetsnummer: "123",
-          erAktiv: false,
-          aktivFom: "2000-02-12T15:24:54.295",
-          aktivTom: null,
-        },
-      },
-      arbeidstaker: {
-        navn: "Labben Rekemester",
-        fnr: "110110110110",
-        epost: null,
-        tlf: null,
-        sistInnlogget: "2020-02-11T15:25:42.275",
-        samtykke: true,
-        evaluering: {
-          effekt: null,
-          hvorfor: null,
-          videre: null,
-          interneaktiviteter: false,
-          ekstratid: false,
-          bistand: false,
-          ingen: false,
-        },
-        stillinger: [
-          {
-            virksomhetsnummer: "110110110",
-            yrke: "Soppkontrollør",
-            prosent: 80,
-            fom: "2000-02-11T15:34:53.787",
-            tom: "2200-02-11T15:34:53.787",
-          },
-        ],
-      },
-      sistEndretAv: {
-        navn: defaultNarmesteLedereMockData[0].navn,
-        fnr: defaultNarmesteLedereMockData[0].fnr,
-        epost: null,
-        tlf: null,
-        sistInnlogget: null,
-        samtykke: null,
-        evaluering: null,
-        stillinger: [],
-      },
-    },
-    {
-      id: 3451,
-      sistEndretDato: "2020-02-20T10:46:06.041",
-      opprettetDato: "2020-02-12",
-      status: "AVBRUTT",
-      virksomhet: {
-        virksomhetsnummer: "110110110",
-        navn: "",
-      },
-      godkjentPlan: {
-        opprettetTidspunkt: "2020-02-20T10:46:04.736",
-        gyldighetstidspunkt: {
-          fom: "2020-02-10",
-          tom: "2020-02-21",
-          evalueres: "2020-02-20",
-        },
-        tvungenGodkjenning: true,
-        deltMedNAVTidspunkt: "2020-02-20T10:46:04.732",
-        deltMedNAV: true,
-        deltMedFastlegeTidspunkt: null,
-        deltMedFastlege: false,
-        dokumentUuid: "f054234f-e989-4d14-81a2-e8c1b1023e4b",
-        avbruttPlan: {
-          tidspunkt: "2020-02-20T11:31:29.835",
-          id: 3451,
-        },
-      },
-      godkjenninger: [],
-      arbeidsoppgaveListe: [
-        {
-          arbeidsoppgaveId: 6397,
-          arbeidsoppgavenavn: "Kontorarbeid ",
-          erVurdertAvSykmeldt: true,
-          gjennomfoering: {
-            kanGjennomfoeres: "TILRETTELEGGING",
-            paaAnnetSted: true,
-            medMerTid: true,
-            medHjelp: false,
-            kanBeskrivelse:
-              "Trenger mer tid til de faste oppgavene og kan utføre noen av disse hjemmefra. ",
-            kanIkkeBeskrivelse: null,
-          },
-          opprettetDato: "2020-02-17T14:37:24.844",
-          sistEndretDato: "2020-02-17T14:37:24.844",
-          sistEndretAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          opprettetAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-        {
-          arbeidsoppgaveId: 6392,
-          arbeidsoppgavenavn: "Utviklingssamtaler",
-          erVurdertAvSykmeldt: true,
-          gjennomfoering: {
-            kanGjennomfoeres: "KAN",
-            paaAnnetSted: null,
-            medMerTid: null,
-            medHjelp: null,
-            kanBeskrivelse: null,
-            kanIkkeBeskrivelse: null,
-          },
-          opprettetDato: "2020-02-17T09:15:03.094",
-          sistEndretDato: "2020-02-17T09:15:03.094",
-          sistEndretAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          opprettetAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-        {
-          arbeidsoppgaveId: 6396,
-          arbeidsoppgavenavn: "Foreldremøte ",
-          erVurdertAvSykmeldt: true,
-          gjennomfoering: {
-            kanGjennomfoeres: "KAN_IKKE",
-            paaAnnetSted: null,
-            medMerTid: null,
-            medHjelp: null,
-            kanBeskrivelse: null,
-            kanIkkeBeskrivelse: "Blir for krevende ",
-          },
-          opprettetDato: "2020-02-17T12:27:03.736",
-          sistEndretDato: "2020-02-17T12:27:22.623",
-          sistEndretAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          opprettetAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-        {
-          arbeidsoppgaveId: 6389,
-          arbeidsoppgavenavn: "Undervisning ",
-          erVurdertAvSykmeldt: false,
-          gjennomfoering: null,
-          opprettetDato: "2020-02-17T09:04:59.149",
-          sistEndretDato: "2020-02-17T09:04:59.149",
-          sistEndretAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          opprettetAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-      ],
-      tiltakListe: [
-        {
-          tiltakId: 4841,
-          tiltaknavn: "Hjemmekontor ",
-          knyttetTilArbeidsoppgaveId: null,
-          fom: "2020-02-10",
-          tom: "2020-02-21",
-          beskrivelse:
-            "Lage og rette prøver, forberede utviklingssamtaler, samt andre enklere oppgaver ",
-          beskrivelseIkkeAktuelt: null,
-          opprettetDato: "2020-02-17T12:53:54.245",
-          sistEndretDato: "2020-02-17T12:53:54.245",
-          kommentarer: [
-            {
-              id: 778,
-              tekst:
-                "Her må vi bli enige om antall timer og hvor mange dager. ",
-              opprettetTidspunkt: "2020-02-17T14:19:02.32",
-              sistEndretDato: "2020-02-17T14:19:02.32",
-              opprettetAv: {
-                navn: defaultNarmesteLedereMockData[0].navn,
-                fnr: defaultNarmesteLedereMockData[0].fnr,
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-              sistEndretAv: {
-                navn: defaultNarmesteLedereMockData[0].navn,
-                fnr: defaultNarmesteLedereMockData[0].fnr,
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-            },
-          ],
-          status: "FORSLAG",
-          gjennomfoering: null,
-          opprettetAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          sistEndretAv: {
-            navn: "Labben Rekemester",
-            fnr: "110110110110",
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-        {
-          tiltakId: 4842,
-          tiltaknavn: "Kortere arbeidsdag",
-          knyttetTilArbeidsoppgaveId: null,
-          fom: "2020-02-10",
-          tom: "2020-02-21",
-          beskrivelse: "Undervise i ett fag for å unngå overbelastning ",
-          beskrivelseIkkeAktuelt: null,
-          opprettetDato: "2020-02-17T14:28:54.574",
-          sistEndretDato: "2020-02-17T14:28:54.574",
-          kommentarer: [
-            {
-              id: 780,
-              tekst: "Hvilket fag passer det best å starte med? ",
-              opprettetTidspunkt: "2020-02-17T14:29:11.568",
-              sistEndretDato: "2020-02-17T14:29:11.568",
-              opprettetAv: {
-                navn: defaultNarmesteLedereMockData[0].navn,
-                fnr: defaultNarmesteLedereMockData[0].fnr,
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-              sistEndretAv: {
-                navn: defaultNarmesteLedereMockData[0].navn,
-                fnr: defaultNarmesteLedereMockData[0].fnr,
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-            },
-            {
-              id: 781,
-              tekst: "Gjerne KRLE eller samfunnsfag.",
-              opprettetTidspunkt: "2020-02-17T14:30:24.019",
-              sistEndretDato: "2020-02-17T14:30:24.019",
-              opprettetAv: {
-                navn: "Labben Rekemester",
-                fnr: "110110110110",
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-              sistEndretAv: {
-                navn: "Labben Rekemester",
-                fnr: "110110110110",
-                epost: null,
-                tlf: null,
-                sistInnlogget: null,
-                samtykke: null,
-                evaluering: null,
-                stillinger: [],
-              },
-            },
-          ],
-          status: "AVTALT",
-          gjennomfoering:
-            "Vi ser hvordan det går denne perioden og øker gradvis hvis aktuelt. ",
-          opprettetAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          sistEndretAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-        {
-          tiltakId: 4843,
-          tiltaknavn: "Assistent i klasserommet ",
-          knyttetTilArbeidsoppgaveId: null,
-          fom: null,
-          tom: null,
-          beskrivelse: "Assistent til å bistå i undervisningen.",
-          beskrivelseIkkeAktuelt:
-            "Skolen har ikke ressurser til å kunne tilby dette. ",
-          opprettetDato: "2020-02-17T14:32:51.421",
-          sistEndretDato: "2020-02-17T14:32:51.421",
-          kommentarer: [],
-          status: "IKKE_AKTUELT",
-          gjennomfoering: null,
-          opprettetAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-          sistEndretAv: {
-            navn: defaultNarmesteLedereMockData[0].navn,
-            fnr: defaultNarmesteLedereMockData[0].fnr,
-            epost: null,
-            tlf: null,
-            sistInnlogget: null,
-            samtykke: null,
-            evaluering: null,
-            stillinger: [],
-          },
-        },
-      ],
-      avbruttPlanListe: [
-        {
-          tidspunkt: "2020-02-12T15:24:53.34",
-          id: 3450,
-        },
-      ],
-      arbeidsgiver: {
-        narmesteLeder: {
-          navn: defaultNarmesteLedereMockData[0].navn,
-          fnr: "20405060718",
-          epost: null,
-          tlf: null,
-          sistInnlogget: "2020-02-20T11:31:31.356",
-          samtykke: true,
-          virksomhetsnummer: "123",
-          erAktiv: false,
-          aktivFom: "2010-02-20T11:31:31.356",
-          aktivTom: null,
-        },
-      },
-      arbeidstaker: {
-        navn: "Labben Rekemester",
-        fnr: "110110110110",
-        epost: null,
-        tlf: null,
-        sistInnlogget: "2020-02-17T14:33:38.179",
-        samtykke: null,
-        evaluering: {
-          effekt: null,
-          hvorfor: null,
-          videre: null,
-          interneaktiviteter: false,
-          ekstratid: false,
-          bistand: false,
-          ingen: false,
-        },
-        stillinger: [
-          {
-            virksomhetsnummer: "110110110",
-            yrke: "Soppkontrollør",
-            prosent: 80,
-            fom: "2000-02-11T15:34:53.787",
-            tom: "2200-02-11T15:34:53.787",
-          },
-        ],
-      },
-      sistEndretAv: {
-        navn: defaultNarmesteLedereMockData[0].navn,
-        fnr: defaultNarmesteLedereMockData[0].fnr,
-        epost: null,
-        tlf: null,
-        sistInnlogget: null,
-        samtykke: null,
-        evaluering: null,
-        stillinger: [],
-      },
-    },
-  ];
