--- src/server/data/mock/testscenarios/godkjentPlanAvslatt/godkjennPlanAvslattScenario.ts
@@ -1,9 +1,0 @@
-import { godkjennPlanAvslattOppfolgingsplanerMockData } from "./godkjennPlanAvslattOppfolgingsplanerMockData";
-import { MockSetup } from "../../getMockDb";
-import { defaultMockSetup } from "../../defaultData/defaultMockSetup";
-
-export const godkjennPlanAvslattScenario: MockSetup = {
-  ...defaultMockSetup,
-  oppfolgingsplaner: godkjennPlanAvslattOppfolgingsplanerMockData,
-  activeTestScenario: "GODKJENNPLANAVSLATT",
-};
