+++ src/server/data/mock/testscenarios/ingenplan/ingenPlanScenario.ts
@@ -1,7 +1,7 @@
 import { MockSetup } from "../../getMockDb";
 import { defaultMockSetup } from "../../defaultData/defaultMockSetup";
 
-export const noPlanScenarioSM: MockSetup = {
+export const ingenPlanScenario: MockSetup = {
   ...defaultMockSetup,
   oppfolgingsplaner: [],
   activeTestScenario: "INGENPLAN",
