+++ src/server/data/mock/testscenarios/underarbeid/planUnderArbeidOppfolgingsplanerMockData.ts
@@ -1,0 +1,418 @@
+import { defaultNarmesteLedereMockData } from "server/data/mock/defaultData/oppfolgingsplanservice/defaultNarmesteLedereMockData";
+import { OppfolgingsplanDTO } from "../../../../../schema/oppfolgingsplanSchema";
+
+export const planUnderArbeidOppfolgingsplanerMockData: OppfolgingsplanDTO[] = [
+  {
+    id: 3455,
+    sistEndretDato: "2020-02-20T11:31:29.837",
+    opprettetDato: "2020-02-20",
+    status: "UNDER_ARBEID",
+    virksomhet: {
+      virksomhetsnummer: "110110110",
+      navn: "",
+    },
+    godkjentPlan: null,
+    godkjenninger: [],
+    arbeidsoppgaveListe: [
+      {
+        arbeidsoppgaveId: 6406,
+        arbeidsoppgavenavn: "Kontorarbeid ",
+        erVurdertAvSykmeldt: true,
+        gjennomfoering: {
+          kanGjennomfoeres: "TILRETTELEGGING",
+          paaAnnetSted: true,
+          medMerTid: true,
+          medHjelp: false,
+          kanBeskrivelse:
+            "Trenger mer tid til de faste oppgavene og kan utføre noen av disse hjemmefra. ",
+          kanIkkeBeskrivelse: null,
+        },
+        opprettetDato: "2020-02-20T11:31:30.305",
+        sistEndretDato: "2020-02-20T11:31:30.305",
+        sistEndretAv: {
+          navn: "Labben Rekemester",
+          fnr: "110110110110",
+          epost: null,
+          tlf: null,
+          sistInnlogget: null,
+          samtykke: null,
+          evaluering: null,
+          stillinger: [],
+        },
+        opprettetAv: {
+          navn: "Labben Krabbeklo",
+          fnr: "110110110110",
+          epost: null,
+          tlf: null,
+          sistInnlogget: null,
+          samtykke: null,
+          evaluering: null,
+          stillinger: [],
+        },
+      },
+      {
+        arbeidsoppgaveId: 6407,
+        arbeidsoppgavenavn: "Utviklingssamtaler",
+        erVurdertAvSykmeldt: true,
+        gjennomfoering: {
+          kanGjennomfoeres: "KAN",
+          paaAnnetSted: null,
+          medMerTid: null,
+          medHjelp: null,
+          kanBeskrivelse: null,
+          kanIkkeBeskrivelse: null,
+        },
+        opprettetDato: "2020-02-20T11:31:30.318",
+        sistEndretDato: "2020-02-20T11:31:30.318",
+        sistEndretAv: {
+          navn: "Labben Rekemester",
+          fnr: "110110110110",
+          epost: null,
+          tlf: null,
+          sistInnlogget: null,
+          samtykke: null,
+          evaluering: null,
+          stillinger: [],
+        },
+        opprettetAv: {
+          navn: "Labben Rekemester",
+          fnr: "110110110110",
+          epost: null,
+          tlf: null,
+          sistInnlogget: null,
+          samtykke: null,
+          evaluering: null,
+          stillinger: [],
+        },
+      },
+      {
+        arbeidsoppgaveId: 6408,
+        arbeidsoppgavenavn: "Foreldremøte ",
+        erVurdertAvSykmeldt: true,
+        gjennomfoering: {
+          kanGjennomfoeres: "KAN_IKKE",
+          paaAnnetSted: null,
+          medMerTid: null,
+          medHjelp: null,
+          kanBeskrivelse: null,
+          kanIkkeBeskrivelse: "Blir for krevende ",
+        },
+        opprettetDato: "2020-02-20T11:31:30.319",
+        sistEndretDato: "2020-02-20T11:31:30.319",
+        sistEndretAv: {
+          navn: defaultNarmesteLedereMockData[0].navn,
+          fnr: defaultNarmesteLedereMockData[0].fnr,
+          epost: null,
+          tlf: null,
+          sistInnlogget: null,
+          samtykke: null,
+          evaluering: null,
+          stillinger: [],
+        },
+        opprettetAv: {
+          navn: defaultNarmesteLedereMockData[0].navn,
+          fnr: defaultNarmesteLedereMockData[0].fnr,
+          epost: null,
+          tlf: null,
+          sistInnlogget: null,
+          samtykke: null,
+          evaluering: null,
+          stillinger: [],
+        },
+      },
+      {
+        arbeidsoppgaveId: 6409,
+        arbeidsoppgavenavn: "Undervisning ",
+        erVurdertAvSykmeldt: false,
+        gjennomfoering: null,
+        opprettetDato: "2020-02-20T11:31:30.322",
+        sistEndretDato: "2020-02-20T11:31:30.322",
+        sistEndretAv: {
+          navn: defaultNarmesteLedereMockData[0].navn,
+          fnr: defaultNarmesteLedereMockData[0].fnr,
+          epost: null,
+          tlf: null,
+          sistInnlogget: null,
+          samtykke: null,
+          evaluering: null,
+          stillinger: [],
+        },
+        opprettetAv: {
+          navn: defaultNarmesteLedereMockData[0].navn,
+          fnr: defaultNarmesteLedereMockData[0].fnr,
+          epost: null,
+          tlf: null,
+          sistInnlogget: null,
+          samtykke: null,
+          evaluering: null,
+          stillinger: [],
+        },
+      },
+    ],
+    tiltakListe: [
+      {
+        tiltakId: 4852,
+        tiltaknavn: "Hjemmekontor ",
+        knyttetTilArbeidsoppgaveId: null,
+        fom: "2021-04-15",
+        tom: "2022-07-26",
+        beskrivelse:
+          "Lage og rette prøver, forberede utviklingssamtaler, samt andre enklere oppgaver ",
+        beskrivelseIkkeAktuelt: null,
+        opprettetDato: "2020-02-20T11:31:30.338",
+        sistEndretDato: "2020-02-20T11:31:30.338",
+        kommentarer: [
+          {
+            id: 782,
+            tekst: "Her må vi bli enige om antall timer og hvor mange dager. ",
+            opprettetTidspunkt: "2020-02-20T11:31:30.352",
+            sistEndretDato: "2020-02-20T11:31:30.352",
+            opprettetAv: {
+              navn: defaultNarmesteLedereMockData[0].navn,
+              fnr: defaultNarmesteLedereMockData[0].fnr,
+              epost: null,
+              tlf: null,
+              sistInnlogget: null,
+              samtykke: null,
+              evaluering: null,
+              stillinger: [],
+            },
+            sistEndretAv: {
+              navn: defaultNarmesteLedereMockData[0].navn,
+              fnr: defaultNarmesteLedereMockData[0].fnr,
+              epost: null,
+              tlf: null,
+              sistInnlogget: null,
+              samtykke: null,
+              evaluering: null,
+              stillinger: [],
+            },
+          },
+        ],
+        status: "FORSLAG",
+        gjennomfoering: null,
+        opprettetAv: {
+          navn: "Labben Rekemester",
+          fnr: "110110110110",
+          epost: null,
+          tlf: null,
+          sistInnlogget: null,
+          samtykke: null,
+          evaluering: null,
+          stillinger: [],
+        },
+        sistEndretAv: {
+          navn: "Labben Rekemester",
+          fnr: "110110110110",
+          epost: null,
+          tlf: null,
+          sistInnlogget: null,
+          samtykke: null,
+          evaluering: null,
+          stillinger: [],
+        },
+      },
+      {
+        tiltakId: 4853,
+        tiltaknavn: "Kortere arbeidsdag",
+        knyttetTilArbeidsoppgaveId: null,
+        fom: "2020-02-10",
+        tom: "2020-02-21",
+        beskrivelse: "Undervise i ett fag for å unngå overbelastning ",
+        beskrivelseIkkeAktuelt: null,
+        opprettetDato: "2020-02-20T11:31:30.373",
+        sistEndretDato: "2020-02-20T11:31:30.373",
+        kommentarer: [
+          {
+            id: 783,
+            tekst: "Hvilket fag passer det best å starte med? ",
+            opprettetTidspunkt: "2020-02-20T11:31:30.376",
+            sistEndretDato: "2020-02-20T11:31:30.376",
+            opprettetAv: {
+              navn: defaultNarmesteLedereMockData[0].navn,
+              fnr: defaultNarmesteLedereMockData[0].fnr,
+              epost: null,
+              tlf: null,
+              sistInnlogget: null,
+              samtykke: null,
+              evaluering: null,
+              stillinger: [],
+            },
+            sistEndretAv: {
+              navn: defaultNarmesteLedereMockData[0].navn,
+              fnr: defaultNarmesteLedereMockData[0].fnr,
+              epost: null,
+              tlf: null,
+              sistInnlogget: null,
+              samtykke: null,
+              evaluering: null,
+              stillinger: [],
+            },
+          },
+          {
+            id: 784,
+            tekst: "Gjerne KRLE eller samfunnsfag.",
+            opprettetTidspunkt: "2020-02-20T11:31:30.378",
+            sistEndretDato: "2020-02-20T11:31:30.378",
+            opprettetAv: {
+              navn: "Labben Rekemester",
+              fnr: "110110110110",
+              epost: null,
+              tlf: null,
+              sistInnlogget: null,
+              samtykke: null,
+              evaluering: null,
+              stillinger: [],
+            },
+            sistEndretAv: {
+              navn: "Labben Rekemester",
+              fnr: "110110110110",
+              epost: null,
+              tlf: null,
+              sistInnlogget: null,
+              samtykke: null,
+              evaluering: null,
+              stillinger: [],
+            },
+          },
+          {
+            id: 785,
+            tekst: "Men, ønsker kun å undervise digitalt.",
+            opprettetTidspunkt: "2020-02-21T12:33:12.378",
+            sistEndretDato: "2020-02-21T12:33:12.378",
+            opprettetAv: {
+              navn: "Labben Rekemester",
+              fnr: "110110110110",
+              epost: null,
+              tlf: null,
+              sistInnlogget: null,
+              samtykke: null,
+              evaluering: null,
+              stillinger: [],
+            },
+            sistEndretAv: {
+              navn: "Labben Rekemester",
+              fnr: "110110110110",
+              epost: null,
+              tlf: null,
+              sistInnlogget: null,
+              samtykke: null,
+              evaluering: null,
+              stillinger: [],
+            },
+          },
+        ],
+        status: "AVTALT",
+        gjennomfoering:
+          "Vi ser hvordan det går denne perioden og øker gradvis hvis aktuelt. ",
+        opprettetAv: {
+          navn: defaultNarmesteLedereMockData[0].navn,
+          fnr: defaultNarmesteLedereMockData[0].fnr,
+          epost: null,
+          tlf: null,
+          sistInnlogget: null,
+          samtykke: null,
+          evaluering: null,
+          stillinger: [],
+        },
+        sistEndretAv: {
+          navn: defaultNarmesteLedereMockData[0].navn,
+          fnr: defaultNarmesteLedereMockData[0].fnr,
+          epost: null,
+          tlf: null,
+          sistInnlogget: null,
+          samtykke: null,
+          evaluering: null,
+          stillinger: [],
+        },
+      },
+      {
+        tiltakId: 4854,
+        tiltaknavn: "Assistent i klasserommet ",
+        knyttetTilArbeidsoppgaveId: null,
+        fom: null,
+        tom: null,
+        beskrivelse: "Assistent til å bistå i undervisningen.",
+        beskrivelseIkkeAktuelt:
+          "Skolen har ikke ressurser til å kunne tilby dette. ",
+        opprettetDato: "2020-02-20T11:31:30.38",
+        sistEndretDato: "2020-02-20T11:31:30.38",
+        kommentarer: [],
+        status: "IKKE_AKTUELT",
+        gjennomfoering: null,
+        opprettetAv: {
+          navn: defaultNarmesteLedereMockData[0].navn,
+          fnr: defaultNarmesteLedereMockData[0].fnr,
+          epost: null,
+          tlf: null,
+          sistInnlogget: null,
+          samtykke: null,
+          evaluering: null,
+          stillinger: [],
+        },
+        sistEndretAv: {
+          navn: defaultNarmesteLedereMockData[0].navn,
+          fnr: defaultNarmesteLedereMockData[0].fnr,
+          epost: null,
+          tlf: null,
+          sistInnlogget: null,
+          samtykke: null,
+          evaluering: null,
+          stillinger: [],
+        },
+      },
+    ],
+    avbruttPlanListe: [
+      {
+        tidspunkt: "2020-02-20T11:31:29.835",
+        id: 3451,
+      },
+      {
+        tidspunkt: "2020-02-12T15:24:53.34",
+        id: 3450,
+      },
+    ],
+    arbeidsgiver: {
+      narmesteLeder: {
+        navn: "Labben Rekemester",
+        fnr: "20405060718",
+        epost: null,
+        tlf: null,
+        sistInnlogget: "2020-02-24T08:57:46.747",
+        samtykke: null,
+        virksomhetsnummer: "123",
+        erAktiv: false,
+        aktivFom: "2020-02-24T08:57:46.747",
+        aktivTom: null,
+      },
+    },
+    arbeidstaker: {
+      navn: "Labben Rekemester",
+      fnr: "110110110110",
+      epost: null,
+      tlf: null,
+      sistInnlogget: "2020-02-24T09:34:24.266",
+      samtykke: null,
+      evaluering: null,
+      stillinger: [
+        {
+          virksomhetsnummer: "110110110",
+          yrke: "Soppkontrollør",
+          prosent: 80,
+          fom: "2000-02-11T15:34:53.787",
+          tom: "2022-02-11T15:34:53.787",
+        },
+      ],
+    },
+    sistEndretAv: {
+      navn: defaultNarmesteLedereMockData[0].navn,
+      fnr: defaultNarmesteLedereMockData[0].fnr,
+      epost: null,
+      tlf: null,
+      sistInnlogget: null,
+      samtykke: null,
+      evaluering: null,
+      stillinger: [],
+    },
+  },
+];
