+++ src/server/data/mock/testscenarios/underarbeid/planUnderArbeidScenario.ts
@@ -1,0 +1,9 @@
+import { planUnderArbeidOppfolgingsplanerMockData } from "./planUnderArbeidOppfolgingsplanerMockData";
+import { MockSetup } from "../../getMockDb";
+import { defaultMockSetup } from "../../defaultData/defaultMockSetup";
+
+export const planUnderArbeidScenario: MockSetup = {
+  ...defaultMockSetup,
+  oppfolgingsplaner: planUnderArbeidOppfolgingsplanerMockData,
+  activeTestScenario: "UNDERARBEID",
+};
